requirements.txt
fastapi
uvicorn
celery[redis]
types-celery
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
python-dotenv   
python-multipart
requests
google-cloud-storage
tqdm